Split amp-date-display component files
